ai/requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
transformers==4.36.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
python-jose==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2